requirements.txt
aiohttp==3.5.4
aioinflux==0.7.0
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
ciso8601==2.1.1
crate==0.23.0
graphitesend==0.10.0
idna==2.8
idna-ssl==1.1.0
influxdb==5.2.2
multidict==4.5.2
numpy==1.16.3
psycopg2-binary==2.8.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
typing-extensions==3.7.2
urllib3==1.24.2
yarl==1.3.0
